Adding pagination wip

+++ src/main/java/api/motortracker/motortracker/repository/CarRepository.java
@@ -3,7 +3,7 @@ package api.motortracker.motortracker.repository;
 import api.motortracker.motortracker.model.Car;
 import org.springframework.data.repository.CrudRepository;
 
-public interface CarRepository extends CrudRepository<Car, Integer> {
+public interface CarRepository extends CrudRepository<Car, Long> {
 
     Car findByPlate(String plate);
 }
