+++ src/main/java/api/motortracker/motortracker/repository/TrackerRepository.java
@@ -3,7 +3,7 @@ package api.motortracker.motortracker.repository;
 import api.motortracker.motortracker.model.Tracker;
 import org.springframework.data.repository.CrudRepository;
 
-public interface TrackerRepository extends CrudRepository<Tracker, Integer> {
+public interface TrackerRepository extends CrudRepository<Tracker, Long> {
 
     Tracker findBySerialNumber(String serialNumber);
 }
